fix: solana block height/confirmations logic
