+++ wormhole-connect/src/sdk/config.ts
@@ -1,5 +1,5 @@
-import { ChainConfig } from '@wormhole-foundation/wormhole-connect-sdk';
-import { CONFIG as CONF } from '@wormhole-foundation/wormhole-connect-sdk';
+import { Network as Environment } from '@certusone/wormhole-sdk';
+import { WormholeContext, ChainConfig, CONFIG as CONF } from '@wormhole-foundation/wormhole-connect-sdk';
 import { MAINNET_NETWORKS, MAINNET_TOKENS } from '../config/mainnet';
 import { TESTNET_NETWORKS, TESTNET_TOKENS } from '../config/testnet';
 import { TokenConfig } from 'config/types';
@@ -14,6 +14,30 @@ console.log('CONFIG', config);
 const { REACT_APP_ENV } = process.env;
 export const isProduction = REACT_APP_ENV === 'MAINNET';
 export const CONFIG = isProduction ? CONF.MAINNET : CONF.TESTNET;
+
+const conf = WormholeContext.getConfig(
+  REACT_APP_ENV! as Environment,
+);
+const mainnetRpcs = {
+  ethereum: process.env.REACT_APP_ETHEREUM_RPC || conf.rpcs.ethereum,
+  solana: process.env.REACT_APP_SOLANA_RPC || conf.rpcs.solana,
+  polygon: process.env.REACT_APP_POLYGON_RPC || conf.rpcs.polygon,
+  bsc: process.env.REACT_APP_BSC_RPC || conf.rpcs.bsc,
+  avalanche: process.env.REACT_APP_AVALANCHE_RPC || conf.rpcs.avalanche,
+  fantom: process.env.REACT_APP_FANTOM_RPC || conf.rpcs.fantom,
+  celo: process.env.REACT_APP_CELO_RPC || conf.rpcs.celo,
+};
+const testnetRpcs = {
+  goerli: process.env.REACT_APP_GOERLI_RPC || conf.rpcs.goerli,
+  mumbai: process.env.REACT_APP_MUMBAI_RPC || conf.rpcs.mumbai,
+  bsc: process.env.REACT_APP_BSC_TESTNET_RPC || conf.rpcs.bsc,
+  fuji: process.env.REACT_APP_FUJI_RPC || conf.rpcs.fuji,
+  fantom: process.env.REACT_APP_FANTOM_TESTNET_RPC || conf.rpcs.fantom,
+  alfajores: process.env.REACT_APP_ALFAJORES_RPC || conf.rpcs.alfajores,
+  solana: process.env.REACT_APP_SOLANA_DEVNET_RPC || conf.rpcs.solana,
+};
+conf.rpcs = REACT_APP_ENV === 'MAINNET' ? mainnetRpcs : testnetRpcs;
+export const WH_CONFIG = config;
 
 export const CHAINS = isProduction ? MAINNET_NETWORKS : TESTNET_NETWORKS;
 export const CHAINS_ARR =
