+++ wormhole-connect/src/utils/index.ts
@@ -5,7 +5,7 @@ import {
   ChainName,
 } from '@wormhole-foundation/wormhole-connect-sdk';
 import { CHAINS_ARR, TOKENS, TOKENS_ARR } from '../sdk/config';
-import { WalletType } from 'store/wallet';
+import { WalletType } from './wallet';
 
 export function convertAddress(address: string): string {
   if (address.length === 22) return address;
